fixing blog put endpoint

+++ controllers/blog.js
@@ -26,8 +26,7 @@ const getAllArticles = async (req, res) => {
       return res.status(200).json(handleResponse('success', 200, { message: "no articles in the database" }))
     }
     return res.status(200).json(handleResponse('success', 200, articles))
-  }
-  catch (error) {
+  } catch (error) {
     res.status(500).json(handleResponse('fail', 500, { message: error.message || 'internal server error' }))
   }
 }
@@ -41,52 +40,44 @@ const updateArticle = async (req, res) => {
     const articleToUpdate = await articleModel.findById(id)
     if (!articleToUpdate) {
       return res.status(404).json(handleResponse("fail", 404, { message: "article not found" }));
-    } else {
-      const updatedArticle = await articleModel.findByIdAndUpdate(id, req.body, { new: true });
-      res.status(200).json(handleResponse("success", 200, updatedArticle))
     }
+    const updatedArticle = await articleModel.findByIdAndUpdate(id, req.body, { new: true });
+    res.status(200).json(handleResponse("success", 200, updatedArticle))
   } catch (error) {
     res.status(500).json(handleResponse('fail', 500, { message: error.message || 'internal server error' }))
   }
-
 }
-
 
 const getSingleArticle = async (req, res) => {
   const id = req.params.articleId
   try {
+    if (id.length != 24) {
+      return res.status(404).json(handleResponse("fail", 404, { message: "article not found" }));
+    }
     const articleExist = await articleModel.findById(id)
-    if (articleExist) {
-      return res.status(200).json(articleExist)
+    if (!articleExist) {
+      return res.status(404).json(handleResponse('fail', 404, { message: 'article with not found' }))
     }
-    else {
-      return res.status(400).json({ "message": `article with id ${id} not found` })
-    }
+    return res.status(200).json(handleResponse('fail', 200, articleExist))
   } catch (error) {
-    res.status(500).json({ "message": error.message })
+    res.status(500).json(handleResponse('fail', 500, { message: error.message || 'internal server error' }))
   }
 }
 
 const deleteArticle = async (req, res) => {
   let id = req.params.articleId
   try {
+    if (id.length != 24) {
+      return res.status(404).json(handleResponse('fail', 404, { message: "article not found" }));
+    }
     const articleFound = await articleModel.findByIdAndDelete(id)
-    if (articleFound) {
-      res.status(200).json({
-        message: `article with id ${id} deleted`
-      })
+    if (!articleFound) {
+      return res.status(404).json(handleResponse('fail', 404, { message: "article not found" }))
     }
-    else {
-      res.status(400).json({
-        message: `article with id ${id} not found`
-      })
-    }
-  } catch (error) {
-    res.status(500).json({
-      "message": error.message
-    })
+    return res.status(200).json(handleResponse('success', 200, { message: "article deleted" }))
+  }
+  catch (error) {
+    return res.status(500).json(handleResponse('fail', 500, { message: error.message || 'internal server error' }))
   }
 }
-
-
 export { createNewArticle, getAllArticles, updateArticle, getSingleArticle, deleteArticle }
